Remove Redundant Code
I have decide that I will be using a webcam for line following as the PiCam throws up to many issues if it is not connected correctly etc.  So in this version I have removed most of the code that dealt with the PiCam.  There is a little setup code left, just in case I change my mind.  The code that I have removed can easily be put back.

I have also taken out most of the code for keyboard control as I have no plans to actually use it.  It's Wiimote all the way.

I also added some sound files so that various actions will cause the bot to make Dalek noises.  Tis has necesitated the addition of an Amp to the bot as the sound to my speaker was far to quiet for a Dalek!  As luck would have it I had a handy, battery powered,  amp looking for somewhere to be useful.