Fixed sound commands
Fixed the lines that play sound and removed some other minor errors.

This one runs.

I also added a -v switch to control sound volume.  But I can't work out how to get it into the command so I have defaulted the volume to 2 so I do not defen the family while testing.